Move sbt plugin project to runners/sbt
