+++ core/src/main/scala/stryker4s/extension/ImplicitMutationConversion.scala
@@ -9,7 +9,7 @@ import scala.meta.Tree
   * Converts [[stryker4s.extension.mutationtype.SubstitutionMutation]] to a `scala.meta.Tree`
   *
   * {{{
-  *  import stryker4s.extensions.ImplicitMutationConversion._
+  *  import stryker4s.extension.ImplicitMutationConversion._
   *  val gt: Tree = GreaterThan
   * }}}
   */
