--- core/src/main/scala/stryker4s/extension/exception/InitialTestRunFailedException.scala
@@ -1,5 +1,0 @@
-package stryker4s.extension.exception
-
-import scala.util.control.NoStackTrace
-
-case class InitialTestRunFailedException(message: String) extends Exception(message) with NoStackTrace
