--- core/src/main/scala/stryker4s/extension/exception/InvalidThresholdValueException.scala
@@ -1,3 +1,0 @@
-package stryker4s.extension.exception
-
-case class InvalidThresholdValueException(message: String) extends Exception(message)
