--- core/src/main/scala/stryker4s/extension/exception/UnableToBuildPatternMatchException.scala
@@ -1,3 +1,0 @@
-package stryker4s.extension.exception
-
-case class UnableToBuildPatternMatchException() extends Exception("Unable to build pattern match")
