+++ core/src/main/scala/stryker4s/extension/mutationtype/MutationTypes.scala
@@ -25,7 +25,7 @@ object Mutation {
 /**
   * Base trait for substitution mutation
   *
-  * Can implicitly be converted to the appropriate `scala.meta.Tree` by importing [[stryker4s.extensions.ImplicitMutationConversion]]
+  * Can implicitly be converted to the appropriate `scala.meta.Tree` by importing [[stryker4s.extension.ImplicitMutationConversion]]
   *
   * @tparam T Has to be a subtype of Tree.
   *           This is so that the tree value and unapply methods return the appropriate type.
