+++ runners/sbt/src/main/scala/stryker4s/sbt/SbtMutantRunner.scala
@@ -5,7 +5,7 @@ import better.files.File
 import sbt._
 import sbt.Keys._
 import stryker4s.config.Config
-import stryker4s.extensions.exceptions.InitialTestRunFailedException
+import stryker4s.extension.exception.InitialTestRunFailedException
 import stryker4s.model._
 import stryker4s.run.MutantRunner
 import stryker4s.run.process.ProcessRunner
