+++ runners/sbt/src/main/scala/stryker4s/sbt/Stryker4sPlugin.scala
