+++ runners/sbt/src/sbt-test/sbt-stryker4s/test-1/project/plugins.sbt
