+++ runners/sbt/src/sbt-test/sbt-stryker4s/test-1/src/main/scala/example/Example.scala
