+++ runners/sbt/src/sbt-test/sbt-stryker4s/test-2.11/project/build.properties
